+++ wp-content/themes/core/blocks/tribe/icon-picker/icons/components/IconAwardCertificate.js
@@ -1,0 +1,32 @@
+const SvgIconAwardCertificate = ( props ) => (
+	<svg
+		width="1em"
+		height="1em"
+		viewBox="0 0 64 64"
+		fill="none"
+		xmlns="http://www.w3.org/2000/svg"
+		{ ...props }
+	>
+		<path
+			d="M32 54H2V2h46v14"
+			stroke="currentColor"
+			strokeWidth={ 2 }
+			strokeMiterlimit={ 10 }
+			strokeLinecap="square"
+		/>
+		<path
+			d="M56 41.486V60l-8-4-8 4V41.486"
+			stroke="currentColor"
+			strokeWidth={ 2 }
+			strokeMiterlimit={ 10 }
+		/>
+		<path
+			d="M48 44c7.732 0 14-6.268 14-14s-6.268-14-14-14-14 6.268-14 14 6.268 14 14 14ZM12 18h16M12 28h12M12 38h12"
+			stroke="currentColor"
+			strokeWidth={ 2 }
+			strokeMiterlimit={ 10 }
+			strokeLinecap="square"
+		/>
+	</svg>
+);
+export default SvgIconAwardCertificate;