+++ wp-content/themes/core/blocks/tribe/icon-picker/icons/components/IconBarChart.js
@@ -1,0 +1,19 @@
+const SvgIconBarChart = ( props ) => (
+	<svg
+		width="1em"
+		height="1em"
+		viewBox="0 0 64 64"
+		fill="none"
+		xmlns="http://www.w3.org/2000/svg"
+		{ ...props }
+	>
+		<path
+			d="M4 58h56M18 26H6v24h12V26ZM38 6H26v44h12V6ZM58 36H46v14h12V36Z"
+			stroke="currentColor"
+			strokeWidth={ 2 }
+			strokeMiterlimit={ 10 }
+			strokeLinecap="square"
+		/>
+	</svg>
+);
+export default SvgIconBarChart;