+++ wp-content/themes/core/blocks/tribe/icon-picker/icons/components/IconBuildingHome.js
@@ -1,0 +1,25 @@
+const SvgIconBuildingHome = ( props ) => (
+	<svg
+		width="1em"
+		height="1em"
+		viewBox="0 0 64 64"
+		fill="none"
+		xmlns="http://www.w3.org/2000/svg"
+		{ ...props }
+	>
+		<path
+			d="M2 28 32 4l30 24M26 60V44h12v16"
+			stroke="currentColor"
+			strokeWidth={ 2 }
+			strokeMiterlimit={ 10 }
+		/>
+		<path
+			d="M10 32v22a6 6 0 0 0 6 6h32a6 6 0 0 0 6-6V32"
+			stroke="currentColor"
+			strokeWidth={ 2 }
+			strokeMiterlimit={ 10 }
+			strokeLinecap="square"
+		/>
+	</svg>
+);
+export default SvgIconBuildingHome;