+++ wp-content/themes/core/blocks/tribe/icon-picker/icons/components/IconCircleConnect.js
@@ -1,0 +1,26 @@
+const SvgIconCircleConnect = ( props ) => (
+	<svg
+		width="1em"
+		height="1em"
+		viewBox="0 0 64 64"
+		fill="none"
+		xmlns="http://www.w3.org/2000/svg"
+		{ ...props }
+	>
+		<path
+			d="M32 14a6 6 0 1 0 0-12 6 6 0 0 0 0 12ZM8.618 54.5a6 6 0 1 0 0-12 6 6 0 0 0 0 12ZM55.382 54.5a6 6 0 1 0 0-12 6 6 0 0 0 0 12Z"
+			stroke="currentColor"
+			strokeWidth={ 2 }
+			strokeMiterlimit={ 10 }
+			strokeLinecap="square"
+		/>
+		<path
+			d="M18.502 11.658A26.974 26.974 0 0 0 6.216 43.034M51.82 53.29a26.894 26.894 0 0 1-33.294 5.088M59 35A27 27 0 0 0 37.932 8.674"
+			stroke="currentColor"
+			strokeWidth={ 2 }
+			strokeMiterlimit={ 10 }
+			strokeLinecap="square"
+		/>
+	</svg>
+);
+export default SvgIconCircleConnect;