+++ wp-content/themes/core/blocks/tribe/icon-picker/icons/components/IconEducationCap.js
@@ -1,0 +1,32 @@
+const SvgIconEducationCap = ( props ) => (
+	<svg
+		width="1em"
+		height="1em"
+		viewBox="0 0 64 64"
+		fill="none"
+		xmlns="http://www.w3.org/2000/svg"
+		{ ...props }
+	>
+		<path
+			d="M32 34 4 20 32 6l28 14-28 14Z"
+			stroke="currentColor"
+			strokeWidth={ 2 }
+			strokeMiterlimit={ 10 }
+			strokeLinecap="square"
+		/>
+		<path
+			d="M12 24v22c0 5.522 8.954 10 20 10s20-4.478 20-10V24"
+			stroke="currentColor"
+			strokeWidth={ 2 }
+			strokeMiterlimit={ 10 }
+		/>
+		<path
+			d="M60 28v14"
+			stroke="currentColor"
+			strokeWidth={ 2 }
+			strokeMiterlimit={ 10 }
+			strokeLinecap="square"
+		/>
+	</svg>
+);
+export default SvgIconEducationCap;