+++ wp-content/themes/core/blocks/tribe/icon-picker/icons/components/IconHeart.js
@@ -1,0 +1,19 @@
+const SvgIconHeart = ( props ) => (
+	<svg
+		width="1em"
+		height="1em"
+		viewBox="0 0 64 64"
+		fill="none"
+		xmlns="http://www.w3.org/2000/svg"
+		{ ...props }
+	>
+		<path
+			d="M55.096 32 32 56 8.904 32A14.213 14.213 0 1 1 32 16a14.214 14.214 0 1 1 23.096 16Z"
+			stroke="currentColor"
+			strokeWidth={ 2 }
+			strokeMiterlimit={ 10 }
+			strokeLinecap="square"
+		/>
+	</svg>
+);
+export default SvgIconHeart;