+++ wp-content/themes/core/blocks/tribe/icon-picker/icons/components/IconHourglassTime.js
@@ -1,0 +1,25 @@
+const SvgIconHourglassTime = ( props ) => (
+	<svg
+		width="1em"
+		height="1em"
+		viewBox="0 0 64 64"
+		fill="none"
+		xmlns="http://www.w3.org/2000/svg"
+		{ ...props }
+	>
+		<path
+			d="M16 60V48c0-10 10-16 10-16s-10-6-10-16V4M48 60V48c0-10-10-16-10-16s10-6 10-16V4"
+			stroke="currentColor"
+			strokeWidth={ 2 }
+			strokeMiterlimit={ 10 }
+		/>
+		<path
+			d="M10 4h44M10 60h44"
+			stroke="currentColor"
+			strokeWidth={ 2 }
+			strokeMiterlimit={ 10 }
+			strokeLinecap="square"
+		/>
+	</svg>
+);
+export default SvgIconHourglassTime;