+++ wp-content/themes/core/blocks/tribe/icon-picker/icons/components/IconInfinity.js
@@ -1,0 +1,26 @@
+const SvgIconInfinity = ( props ) => (
+	<svg
+		width="1em"
+		height="1em"
+		viewBox="0 0 64 64"
+		fill="none"
+		xmlns="http://www.w3.org/2000/svg"
+		{ ...props }
+	>
+		<path
+			d="M41.465 38.875C44.968 41.098 48.595 43 51.09 43 57.091 43 62 38.05 62 32s-4.909-11-10.909-11c-12.182 0-26 22-38.182 22C6.91 43 2 38.05 2 32s4.91-11 10.91-11c4.188 0 8.055 3.025 11.633 5.132.065.039-.163-.095-.098-.058"
+			stroke="currentColor"
+			strokeWidth={ 2 }
+			strokeMiterlimit={ 10 }
+			strokeLinecap="square"
+		/>
+		<path
+			d="M14.375 29 25 26.153l-2.847-10.625"
+			stroke="currentColor"
+			strokeWidth={ 2 }
+			strokeMiterlimit={ 10 }
+			strokeLinecap="square"
+		/>
+	</svg>
+);
+export default SvgIconInfinity;