+++ wp-content/themes/core/blocks/tribe/icon-picker/icons/components/IconMegaphoneSpeaker.js
@@ -1,0 +1,32 @@
+const SvgIconMegaphoneSpeaker = ( props ) => (
+	<svg
+		width="1em"
+		height="1em"
+		viewBox="0 0 64 64"
+		fill="none"
+		xmlns="http://www.w3.org/2000/svg"
+		{ ...props }
+	>
+		<path
+			d="M54 30h8M50.516 17l6.928-4M50.516 43l6.928 4"
+			stroke="currentColor"
+			strokeWidth={ 2 }
+			strokeMiterlimit={ 10 }
+			strokeLinecap="square"
+		/>
+		<path
+			d="m24.294 44.036 5.302 11.882a4.383 4.383 0 0 1-8.004 3.572L11 36"
+			stroke="currentColor"
+			strokeWidth={ 2 }
+			strokeMiterlimit={ 10 }
+		/>
+		<path
+			d="M42 50 22 36H11a9 9 0 0 1 0-18h11L42 4v46ZM14 18v18"
+			stroke="currentColor"
+			strokeWidth={ 2 }
+			strokeMiterlimit={ 10 }
+			strokeLinecap="square"
+		/>
+	</svg>
+);
+export default SvgIconMegaphoneSpeaker;