+++ wp-content/themes/core/blocks/tribe/icon-picker/icons/components/IconSchoolSupplies.js
@@ -1,0 +1,32 @@
+const SvgIconSchoolSupplies = ( props ) => (
+	<svg
+		width="1em"
+		height="1em"
+		viewBox="0 0 64 64"
+		fill="none"
+		xmlns="http://www.w3.org/2000/svg"
+		{ ...props }
+	>
+		<path
+			d="m46 56 12-12M24 10 12 22"
+			stroke="currentColor"
+			strokeWidth={ 2 }
+			strokeMiterlimit={ 10 }
+		/>
+		<path
+			d="M19 29 6 16 18 4l13 13M45 31l13 13 2 14-14-2-13-13"
+			stroke="currentColor"
+			strokeWidth={ 2 }
+			strokeMiterlimit={ 10 }
+			strokeLinecap="square"
+		/>
+		<path
+			d="m43.999 4-40 40 14 13.998 40-39.999L43.998 4ZM24 24l6 6M18 30l4 4M36 12l6 6M30 18l4 4M12 36l6 6"
+			stroke="currentColor"
+			strokeWidth={ 2 }
+			strokeMiterlimit={ 10 }
+			strokeLinecap="square"
+		/>
+	</svg>
+);
+export default SvgIconSchoolSupplies;