+++ wp-content/themes/core/blocks/tribe/icon-picker/icons/components/IconSoundwave.js
@@ -1,0 +1,19 @@
+const SvgIconSoundwave = ( props ) => (
+	<svg
+		width="1em"
+		height="1em"
+		viewBox="0 0 64 64"
+		fill="none"
+		xmlns="http://www.w3.org/2000/svg"
+		{ ...props }
+	>
+		<path
+			d="M28 6v48M36 10v48M12 24v16M20 18v28M4 30v4M60 30v4M44 18v28M52 26v12"
+			stroke="currentColor"
+			strokeWidth={ 2 }
+			strokeMiterlimit={ 10 }
+			strokeLinecap="square"
+		/>
+	</svg>
+);
+export default SvgIconSoundwave;