+++ wp-content/themes/core/blocks/tribe/icon-picker/icons/components/IconStackLayers.js
@@ -1,0 +1,26 @@
+const SvgIconStackLayers = ( props ) => (
+	<svg
+		width="1em"
+		height="1em"
+		viewBox="0 0 64 64"
+		fill="none"
+		xmlns="http://www.w3.org/2000/svg"
+		{ ...props }
+	>
+		<path
+			d="M56.498 30 60 32 32 48 4 32l3.5-2"
+			stroke="currentColor"
+			strokeWidth={ 2 }
+			strokeMiterlimit={ 10 }
+			strokeLinecap="square"
+		/>
+		<path
+			d="M56.51 42.006 60 44 32 60 4 44l3.5-2M4 20 32 4l28 16-28 16L4 20Z"
+			stroke="currentColor"
+			strokeWidth={ 2 }
+			strokeMiterlimit={ 10 }
+			strokeLinecap="square"
+		/>
+	</svg>
+);
+export default SvgIconStackLayers;