+++ wp-content/themes/core/blocks/tribe/icon-picker/icons/components/IconStatistics.js
@@ -1,0 +1,19 @@
+const SvgIconStatistics = ( props ) => (
+	<svg
+		width="1em"
+		height="1em"
+		viewBox="0 0 64 64"
+		fill="none"
+		xmlns="http://www.w3.org/2000/svg"
+		{ ...props }
+	>
+		<path
+			d="M7.998 58 8 44M23.998 26v32M40.036 58 40 44M55.998 26v32M6 26 24 8l16 16L58 6"
+			stroke="currentColor"
+			strokeWidth={ 2 }
+			strokeMiterlimit={ 10 }
+			strokeLinecap="square"
+		/>
+	</svg>
+);
+export default SvgIconStatistics;